+++ ext/src/bg/analysis.js
@@ -1,6 +1,5 @@
-chrome.runtime.onMessage.addListener(
-  function(pgn, sender, sendResponse){
+chrome.runtime.onMessage.addListener(async function(pgn, sender, sendResponse){
     $("textarea[name='pgn']").text(pgn);
     $("input[name='analyse']").prop("checked", true);
-	  $(".submit")[0].click();
+    $(".submit")[0].click();
 });